Move HashableDict class to `utils.py`
