+++ code/diamm/utils.py
@@ -1,5 +1,5 @@
 """
-Module containing utility classes for the async fetching script for the DIAMM database.
+Module containing utility classes for various scripts for the DIAMM database.
 """
 
 import asyncio
@@ -95,3 +95,40 @@ class NotifyingQueue(asyncio.Queue):
         else:
             self._event.clear()
         await self._event.wait()
+
+
+class HashableDict:
+    """
+    A hashable dictionary class that allows for easy comparison and hashing.
+    This class is used to store dictionaries in a set, allowing for easy
+    removal of duplicate relations.
+
+    It is initialized with a dictionary and provides methods for hashing,
+    equality comparison, and representation.
+
+    It also includes a static method to convert an iterable of HashableDict objects
+    to a list of dictionaries, which is useful for converting the set of relations
+    to a DataFrame.
+    """
+
+    def __init__(self, d):
+        self.d = dict(d)
+        self._frozen = frozenset(self.d.items())
+
+    def __hash__(self):
+        return hash(self._frozen)
+
+    def __eq__(self, other):
+        return isinstance(other, HashableDict) and self._frozen == other._frozen
+
+    def __repr__(self):
+        return f"HashableDict({self.d})"
+
+    @staticmethod
+    def to_list_of_dicts(lst):
+        """
+        Takes an iterable object (set, list, etc) containing HashableDict
+        objects and returns a list of dictionaries.
+        This is useful for converting the set of relations to a DataFrame.
+        """
+        return [d.d for d in lst]
